Refactoring error related classes
