+++ lib/groonga/client/error.rb
@@ -18,5 +18,35 @@ module Groonga
   class Client
     class Error < StandardError
     end
+
+    class ErrorResponse < Error
+      attr_reader :response
+      def initialize(response)
+        @response = response
+        command = @response.command
+        status_code = @response.status_code
+        error_message = @response.error_message
+        message = "failed to execute: "
+        message << "#{command.command_name}: #{status_code}: "
+        message << "<#{error_message}>: "
+        message << command.to_command_format
+        super(message)
+      end
+    end
+
+    class InvalidResponse < Client::Error
+      attr_reader :command
+      attr_reader :raw_response
+      def initialize(command, raw_response, error_message)
+        @command = command
+        @raw_response = raw_response
+        message = +"invalid response: "
+        message << "#{command.command_name}: "
+        message << "#{error_message}: "
+        message << "<#{command.to_command_format}>: "
+        message << "<#{raw_response}>"
+        super(message)
+      end
+    end
   end
 end
