+++ test/response/test-base.rb
@@ -226,11 +226,11 @@ class TestResponseBase < Test::Unit::TestCase
     def test_invalid_json
       command = Groonga::Command::Base.new("cancel")
       raw_response = '["header", :{"return_code":-77}}'
-      err = assert_raise(Groonga::Client::Response::InvalidResponse) do
+      err = assert_raise(Groonga::Client::InvalidResponse) do
         Groonga::Client::Response::Base.parse(command, raw_response)
       end
       error_message = <<~'MESSAGE'.chomp
-        invalid response: cancel: invalid JSON: unexpected token at ':{"return_code":-77}}': cancel: <["header", :{"return_code":-77}}>
+        invalid response: cancel: invalid JSON: unexpected token at ':{"return_code":-77}}': <cancel>: <["header", :{"return_code":-77}}>
       MESSAGE
       assert_equal(error_message, err.message)
       assert_equal(command, err.command)
